+++ abstracts/SimpleHash.php
@@ -5,7 +5,7 @@
  */
 namespace HASH\abstracts;
 use HASH\interfaces\iHash;
-interface_exists('iHash') or require 'HASH/interfaces/iHash.php';
+interface_exists('HASH\interfaces\iHash', false) or require 'HASH/interfaces/iHash.php';
 /**
  * @package HASH.abstracts
  * @since 1.0