+++ strategies/Blowfish.php
@@ -5,7 +5,7 @@
  */
 namespace HASH\strategies;
 use HASH\abstracts\SimpleBlowfish;
-class_exists('SimpleBlowfish') or require 'HASH/abstracts/SimpleBlowfish.php';
+class_exists('HASH\abstracts\SimpleBlowfish', false) or require 'HASH/abstracts/SimpleBlowfish.php';
 /**
  * @package HASH.strategies
  * @since 1.1