+++ strategies/BlowfishSalt.php
@@ -6,7 +6,7 @@
 namespace HASH\strategies;
 use HASH\abstracts\SimpleBlowfish;
 use \Exception;
-class_exists('SimpleBlowfish') or require 'HASH/abstracts/SimpleBlowfish.php';
+class_exists('HASH\abstracts\SimpleBlowfish', false) or require 'HASH/abstracts/SimpleBlowfish.php';
 /**
  * @package HASH.strategies
  * @since 1.1
@@ -20,7 +20,7 @@ final class BlowfishSalt extends SimpleBlowfish
 	 */
 	public function __construct(array $config)
 	{
-		if ( ! isset($config['salt']) or ! preg_match('^[./0-9A-Z-a-z]{22}$', $config['salt'])) {
+		if ( ! isset($config['salt']) or ! preg_match('/^[.\/0-9A-Z-a-z]{22}$/', $config['salt'])) {
 			throw new Exception('_invalid_salt');
 		}
 		parent::__construct($config);